refactor: move retries to promise utilities
