+++ src/internal/embedding.ts
@@ -10,7 +10,7 @@
  * private `_localRetriever` config — never in serializable config sent to the browser.
  */
 
-import * as WithRetry from './with-retry.js'
+import * as promise from './promise.js'
 
 /** A single embedding vector. */
 export type Vector = readonly number[]
@@ -290,7 +290,7 @@ export function cloudflare(options: cloudflare.Options = {}): Adapter {
         )
       const p = prefix?.[context.purpose] ?? ''
       const url = `${baseUrl.replace(/\/$/, '')}/accounts/${accountId}/ai/run/${model}`
-      const response = await WithRetry.withRetry(
+      const response = await promise.withRetry(
         () =>
           fetch(url, {
             method: 'POST',
